fixed the same bug for the other tab

+++ src/tab1.rs
@@ -4395,7 +4395,7 @@ impl Tab {
                 }
 
                 count += 1;
-                y += row_height;
+                y += row_height as u32;
                 children.push(button_row);
             }
